+++ Topos/Consumer/IPositionManager.cs
@@ -5,6 +5,6 @@ namespace Topos.Consumer
     public interface IPositionManager
     {
         Task Set(Position position);
-        Task<Position?> Get(string topic, int partition);
+        Task<Position> Get(string topic, int partition);
     }
 }